+++ lib/paper_trail/events/create.rb
@@ -13,6 +13,7 @@ module PaperTrail
       # @api private
       def data
         data = {
+          item: @record,
           event: @record.paper_trail_event || "create",
           whodunnit: PaperTrail.request.whodunnit
         }