+++ lib/paper_trail/version_number.rb
@@ -8,8 +8,8 @@ module PaperTrail
   # People are encouraged to use `PaperTrail.gem_version` instead.
   module VERSION
     MAJOR = 10
-    MINOR = 1
-    TINY = 0
+    MINOR = 3
+    TINY = 1
 
     # Set PRE to nil unless it's a pre-release (beta, rc, etc.)
     PRE = nil