+++ spec/controllers/widgets_controller_spec.rb
@@ -10,7 +10,7 @@ RSpec.describe WidgetsController, type: :controller, versioning: true do
   describe "#create" do
     context "PT enabled" do
       it "stores information like IP address in version" do
-        post(:create, params_wrapper(widget: { name: "Flugel" }))
+        post(:create, params: { widget: { name: "Flugel" } })
         widget = assigns(:widget)
         expect(widget.versions.length).to(eq(1))
         expect(widget.versions.last.whodunnit.to_i).to(eq(153))
@@ -20,7 +20,7 @@ RSpec.describe WidgetsController, type: :controller, versioning: true do
 
       it "controller metadata methods should get evaluated" do
         request.env["HTTP_USER_AGENT"] = "User-Agent"
-        post :create, params_wrapper(widget: { name: "Flugel" })
+        post :create, params: { widget: { name: "Flugel" } }
         expect(PaperTrail.request.enabled?).to eq(true)
         expect(PaperTrail.request.whodunnit).to(eq(153))
         expect(PaperTrail.request.controller_info.present?).to(eq(true))
@@ -32,7 +32,7 @@ RSpec.describe WidgetsController, type: :controller, versioning: true do
     context "PT disabled" do
       it "does not save a version, and metadata is not set" do
         request.env["HTTP_USER_AGENT"] = "Disable User-Agent"
-        post :create, params_wrapper(widget: { name: "Flugel" })
+        post :create, params: { widget: { name: "Flugel" } }
         expect(assigns(:widget).versions.length).to(eq(0))
         expect(PaperTrail.request.enabled?).to eq(false)
         expect(PaperTrail.request.whodunnit).to be_nil
@@ -44,17 +44,17 @@ RSpec.describe WidgetsController, type: :controller, versioning: true do
   describe "#destroy" do
     it "can be disabled" do
       request.env["HTTP_USER_AGENT"] = "Disable User-Agent"
-      post(:create, params_wrapper(widget: { name: "Flugel" }))
+      post(:create, params: { widget: { name: "Flugel" } })
       w = assigns(:widget)
       expect(w.versions.length).to(eq(0))
-      delete(:destroy, params_wrapper(id: w.id))
+      delete(:destroy, params: { id: w.id })
       expect(PaperTrail::Version.with_item_keys("Widget", w.id).size).to(eq(0))
     end
 
     it "stores information like IP address in version" do
       w = Widget.create(name: "Roundel")
       expect(w.versions.length).to(eq(1))
-      delete(:destroy, params_wrapper(id: w.id))
+      delete(:destroy, params: { id: w.id })
       widget = assigns(:widget)
       expect(widget.versions.length).to(eq(2))
       expect(widget.versions.last.ip).to(eq("127.0.0.1"))
@@ -67,7 +67,7 @@ RSpec.describe WidgetsController, type: :controller, versioning: true do
     it "stores information like IP address in version" do
       w = Widget.create(name: "Duvel")
       expect(w.versions.length).to(eq(1))
-      put(:update, params_wrapper(id: w.id, widget: { name: "Bugle" }))
+      put(:update, params: { id: w.id, widget: { name: "Bugle" } })
       widget = assigns(:widget)
       expect(widget.versions.length).to(eq(2))
       expect(widget.versions.last.whodunnit.to_i).to(eq(153))
@@ -77,10 +77,10 @@ RSpec.describe WidgetsController, type: :controller, versioning: true do
 
     it "can be disabled" do
       request.env["HTTP_USER_AGENT"] = "Disable User-Agent"
-      post(:create, params_wrapper(widget: { name: "Flugel" }))
+      post(:create, params: { widget: { name: "Flugel" } })
       w = assigns(:widget)
       expect(w.versions.length).to(eq(0))
-      put(:update, params_wrapper(id: w.id, widget: { name: "Bugle" }))
+      put(:update, params: { id: w.id, widget: { name: "Bugle" } })
       widget = assigns(:widget)
       expect(widget.versions.length).to(eq(0))
     end