+++ spec/dummy_app/db/migrate/20110208155312_set_up_test_tables.rb
@@ -6,13 +6,7 @@
 # Starting with AR 5.1, we must specify which version of AR we are using.
 # I tried using `const_get` but I got a `NameError`, then I learned about
 # `::ActiveRecord::Migration::Current`.
-class SetUpTestTables < (
-  if ::ActiveRecord::VERSION::MAJOR >= 5
-    ::ActiveRecord::Migration::Current
-  else
-    ::ActiveRecord::Migration
-  end
-)
+class SetUpTestTables < ::ActiveRecord::Migration::Current
   MYSQL_ADAPTERS = [
     "ActiveRecord::ConnectionAdapters::MysqlAdapter",
     "ActiveRecord::ConnectionAdapters::Mysql2Adapter"
@@ -78,9 +72,9 @@ class SetUpTestTables < (
     end
 
     create_table :versions, versions_table_options do |t|
-      t.string   :item_type, item_type_options(null: false)
+      t.string   :item_type, **item_type_options(null: false)
       t.integer  :item_id, null: false
-      t.string   :item_subtype, item_type_options(null: true)
+      t.string   :item_subtype, **item_type_options(null: true)
       t.string   :event, null: false
       t.string   :whodunnit
       t.text     :object, limit: TEXT_BYTES
@@ -100,16 +94,6 @@ class SetUpTestTables < (
       t.string :user_agent
     end
     add_index :versions, %i[item_type item_id]
-
-    create_table :version_associations do |t|
-      t.integer  :version_id
-      t.string   :foreign_key_name, null: false
-      t.integer  :foreign_key_id
-    end
-    add_index :version_associations, [:version_id]
-    add_index :version_associations,
-      %i[foreign_key_name foreign_key_id],
-      name: "index_version_associations_on_foreign_key"
 
     create_table :post_versions, force: true do |t|
       t.string   :item_type, null: false